+++ mindformers/tools/dataset_preprocess/pangualpha/cmrc2018_data_process.py
@@ -145,9 +145,9 @@ def write_instance_to_file(writer, instance, need_del_keys):
     label = instance["labels"]
 
     features = collections.OrderedDict()
-    features["input_ids"] = np.asarray(input_ids)
-    features["attention_mask"] = np.asarray(attention_mask)
-    features["labels"] = np.asarray(label)
+    features["input_ids"] = np.asarray(input_ids).astype(np.int32)
+    features["attention_mask"] = np.asarray(attention_mask).astype(np.int32)
+    features["labels"] = np.asarray(label).astype(np.int32)
 
     for need_del_key in need_del_keys:
         del features[need_del_key]
